fixed tests making http requests

+++ server/src/main/publickey_interface/src/containers/KeyInterface.js
@@ -5,7 +5,6 @@ import KJUR, {KEYUTIL} from 'jsrsasign';
 import KeyEntry from '../components/KeyEntry';
 import EditEntry from '../components/EditEntry';
 
-
 export default class KeyInterface extends Component{
     constructor(props){
         super(props);
@@ -28,17 +27,20 @@ export default class KeyInterface extends Component{
     }
 
     async componentDidMount(){
-        var jwtData = await fetch('http://localhost:8090/api/public', {
-            method: 'GET',
-            headers: {
-                'Accept': 'application/json'
-            }
-            }).then(response=>{
-                return response.json();
-            }).catch(error=>{
-                console.log("Could not load data, make sure the server is running.")
-            });
-        this.initData(jwtData);
+        if(this.props.doFetch){
+            var jwtData = await fetch('https://localhost:8090/api/public', {
+                method: 'GET',
+                headers: {
+                    'Accept': 'application/json'
+                }
+                }).then(response=>{
+                    return response.json();
+                }).catch(error=>{
+                    console.log("Could not load data, make sure the server is running.")
+                });
+            this.initData(jwtData);
+        }
+
     }
     async saveData(keyObject){
         // const jwtData = this.state.jwtJson;
@@ -52,49 +54,55 @@ export default class KeyInterface extends Component{
         const keyId = Object.keys(keyObject)[0];
         const key = keyObject[keyId];
         const result = {"id":keyId,"key":key};
-        await fetch('http://localhost:8090/api/public', {
-            method: 'POST',
-            headers: {
-                'Accept': 'application/json',
-                'Contetn': 'application/json'
-            },
-            body: JSON.stringify(result)
-            }).then(response=>{
-                console.log("Saved the data")
-            }).catch(error=>{
-                console.log("Could not save data");
-            });
+        if(this.props.doFetch){
+            await fetch('https://localhost:8090/api/public', {
+                method: 'POST',
+                headers: {
+                    'Accept': 'application/json',
+                    'Contetn': 'application/json'
+                },
+                body: JSON.stringify(result)
+                }).then(response=>{
+                    console.log("Saved the data")
+                }).catch(error=>{
+                    console.log("Could not save data");
+                });
+        }
     }
 
     async deleteData(id){
-        await fetch('http://localhost:8090/api/public/'+id, {
-            method: 'DELETE',
-            headers: {
-                'Accept': 'application/json',
-                'Contetn': 'application/json'
-            }
-            }).then(response=>{
-                console.log("Deleted the data")
-            }).catch(error=>{
-                console.log("Could not save data");
-            });;
+        if(this.props.doFetch){
+            await fetch('https://localhost:8090/api/public/'+id, {
+                method: 'DELETE',
+                headers: {
+                    'Accept': 'application/json',
+                    'Contetn': 'application/json'
+                }
+                }).then(response=>{
+                    console.log("Deleted the data")
+                }).catch(error=>{
+                    console.log("Could not save data");
+                });;
+        }
     }
     async editData(oldId,keyObject){
         const keyId = Object.keys(keyObject)[0];
         const key = keyObject[keyId];
         const result = {"id":keyId,"key":JSON.stringify(key)};
-        await fetch('http://localhost:8090/api/public/'+oldId, {
-            method: 'PUT',
-            headers: {
-                'Accept': 'application/json',
-                'Contetn': 'application/json'
-            },
-            body: JSON.stringify(result)
-            }).then(response=>{
-                console.log("Saved the data")
-            }).catch(error=>{
-                console.log("Could not save data");
-            });;
+        if(this.props.doFetch){
+            await fetch('https://localhost:8090/api/public/'+oldId, {
+                method: 'PUT',
+                headers: {
+                    'Accept': 'application/json',
+                    'Contetn': 'application/json'
+                },
+                body: JSON.stringify(result)
+                }).then(response=>{
+                    console.log("Saved the data")
+                }).catch(error=>{
+                    console.log("Could not save data");
+                });;
+        }
     }
     initData(jwtData){
         if(jwtData){
